+++ Source/WebCore/css/parser/CSSPropertyParserHelpers.cpp
@@ -65,11 +65,6 @@
 #include "CSSCursorImageValue.h"
 #include "CSSCustomPropertyValue.h"
 #include "CSSFunctionValue.h"
-#include "CSSGridAutoRepeatValue.h"
-#include "CSSGridIntegerRepeatValue.h"
-#include "CSSGridLineNamesValue.h"
-#include "CSSGridLineValue.h"
-#include "CSSGridTemplateAreasValue.h"
 #include "CSSLineBoxContainValue.h"
 #include "CSSOffsetRotateValue.h"
 #include "CSSParser.h"
@@ -81,7 +76,6 @@
 #include "CSSRectValue.h"
 #include "CSSReflectValue.h"
 #include "CSSShapeSegmentValue.h"
-#include "CSSSubgridValue.h"
 #include "CSSTransformListValue.h"
 #include "CSSValueKeywords.h"
 #include "CSSValuePair.h"
@@ -2150,48 +2144,6 @@ template<CSSPropertyID property> RefPtr<CSSValue> consumeBackgroundSize(CSSParse
     return CSSValuePair::createNoncoalescing(horizontal.releaseNonNull(), vertical.releaseNonNull());
 }
 
-RefPtr<CSSValue> consumeGridAutoFlow(CSSParserTokenRange& range)
-{
-    auto rowOrColumnValue = consumeIdent<CSSValueRow, CSSValueColumn>(range);
-    auto denseAlgorithm = consumeIdent<CSSValueDense>(range);
-    if (!rowOrColumnValue) {
-        rowOrColumnValue = consumeIdent<CSSValueRow, CSSValueColumn>(range);
-        if (!rowOrColumnValue && !denseAlgorithm)
-            return nullptr;
-    }
-    CSSValueListBuilder parsedValues;
-    if (rowOrColumnValue) {
-        CSSValueID value = rowOrColumnValue->valueID();
-        if (value == CSSValueColumn || (value == CSSValueRow && !denseAlgorithm))
-            parsedValues.append(rowOrColumnValue.releaseNonNull());
-    }
-    if (denseAlgorithm)
-        parsedValues.append(denseAlgorithm.releaseNonNull());
-    return CSSValueList::createSpaceSeparated(WTFMove(parsedValues));
-}
-
-RefPtr<CSSValueList> consumeMasonryAutoFlow(CSSParserTokenRange& range)
-{
-    auto packOrNextValue = consumeIdent<CSSValuePack, CSSValueNext>(range);
-    auto definiteFirstOrOrderedValue = consumeIdent<CSSValueDefiniteFirst, CSSValueOrdered>(range);
-
-    if (!packOrNextValue) {
-        packOrNextValue = consumeIdent<CSSValuePack, CSSValueNext>(range);
-        if (!packOrNextValue)
-            packOrNextValue = CSSPrimitiveValue::create(CSSValuePack);
-    }
-
-    CSSValueListBuilder parsedValues;
-    if (packOrNextValue) {
-        CSSValueID packOrNextValueID = packOrNextValue->valueID();
-        if (!definiteFirstOrOrderedValue || (definiteFirstOrOrderedValue && definiteFirstOrOrderedValue->valueID() == CSSValueDefiniteFirst) || packOrNextValueID == CSSValueNext)
-            parsedValues.append(packOrNextValue.releaseNonNull());
-    }
-    if (definiteFirstOrOrderedValue && definiteFirstOrOrderedValue->valueID() == CSSValueOrdered)
-        parsedValues.append(definiteFirstOrOrderedValue.releaseNonNull());
-    return CSSValueList::createSpaceSeparated(WTFMove(parsedValues));
-}
-
 RefPtr<CSSValue> consumeRepeatStyle(CSSParserTokenRange& range, const CSSParserContext&)
 {
     // https://www.w3.org/TR/css-backgrounds-3/#typedef-repeat-style
@@ -2264,10 +2216,6 @@ bool isSelfPositionOrLeftOrRightKeyword(CSSValueID id)
     return isSelfPositionKeyword(id) || isLeftOrRightKeyword(id);
 }
 
-bool isGridBreadthIdent(CSSValueID id)
-{
-    return identMatches<CSSValueMinContent, CSSValueWebkitMinContent, CSSValueMaxContent, CSSValueWebkitMaxContent, CSSValueAuto>(id);
-}
 
 RefPtr<CSSValue> consumeSelfPositionOverflowPosition(CSSParserTokenRange& range, IsPositionKeyword isPositionKeyword)
 {
@@ -2313,395 +2261,6 @@ RefPtr<CSSValue> consumeJustifyItems(CSSParserTokenRange& range)
     return consumeSelfPositionOverflowPosition(range, isSelfPositionOrLeftOrRightKeyword);
 }
 
-static RefPtr<CSSValue> consumeFitContent(CSSParserTokenRange& range, CSSParserMode mode)
-{
-    CSSParserTokenRange rangeCopy = range;
-    CSSParserTokenRange args = consumeFunction(rangeCopy);
-    auto length = consumeLengthOrPercent(args, mode, ValueRange::NonNegative, UnitlessQuirk::Allow);
-    if (!length || !args.atEnd())
-        return nullptr;
-    range = rangeCopy;
-    return CSSFunctionValue::create(CSSValueFitContent, length.releaseNonNull());
-}
-
-static RefPtr<CSSPrimitiveValue> consumeCustomIdentForGridLine(CSSParserTokenRange& range)
-{
-    if (range.peek().id() == CSSValueAuto || range.peek().id() == CSSValueSpan)
-        return nullptr;
-    return consumeCustomIdent(range);
-}
-
-RefPtr<CSSValue> consumeGridLine(CSSParserTokenRange& range)
-{
-    if (range.peek().id() == CSSValueAuto)
-        return consumeIdent(range);
-
-    RefPtr<CSSPrimitiveValue> spanValue;
-    RefPtr<CSSPrimitiveValue> gridLineName;
-    RefPtr<CSSPrimitiveValue> numericValue = consumeInteger(range);
-    if (numericValue) {
-        gridLineName = consumeCustomIdentForGridLine(range);
-        spanValue = consumeIdent<CSSValueSpan>(range);
-    } else {
-        spanValue = consumeIdent<CSSValueSpan>(range);
-        if (spanValue) {
-            numericValue = consumeInteger(range);
-            gridLineName = consumeCustomIdentForGridLine(range);
-            if (!numericValue)
-                numericValue = consumeInteger(range);
-        } else {
-            gridLineName = consumeCustomIdentForGridLine(range);
-            if (gridLineName) {
-                numericValue = consumeInteger(range);
-                spanValue = consumeIdent<CSSValueSpan>(range);
-                if (!spanValue && !numericValue)
-                    return gridLineName;
-            } else
-                return nullptr;
-        }
-    }
-
-    if (spanValue && !numericValue && !gridLineName)
-        return nullptr; // "span" keyword alone is invalid.
-    if (spanValue && numericValue && numericValue->isNegative().value_or(false))
-        return nullptr; // Negative numbers are not allowed for span.
-    if (numericValue && numericValue->isZero().value_or(false))
-        return nullptr; // An <integer> value of zero makes the declaration invalid.
-
-    return CSSGridLineValue::create(WTFMove(spanValue), WTFMove(numericValue), WTFMove(gridLineName));
-}
-
-static bool isGridTrackFixedSized(const CSSPrimitiveValue& primitiveValue)
-{
-    switch (primitiveValue.valueID()) {
-    case CSSValueMinContent:
-    case CSSValueWebkitMinContent:
-    case CSSValueMaxContent:
-    case CSSValueWebkitMaxContent:
-    case CSSValueAuto:
-        return false;
-    default:
-        return !primitiveValue.isFlex();
-    }
-}
-
-static bool isGridTrackFixedSized(const CSSValue& value)
-{
-    if (auto* primitiveValue = dynamicDowncast<CSSPrimitiveValue>(value))
-        return isGridTrackFixedSized(*primitiveValue);
-    auto& function = downcast<CSSFunctionValue>(value);
-    if (function.name() == CSSValueFitContent || function.length() < 2)
-        return false;
-    return isGridTrackFixedSized(downcast<CSSPrimitiveValue>(*function.item(0)))
-        || isGridTrackFixedSized(downcast<CSSPrimitiveValue>(*function.item(1)));
-}
-
-static Vector<String> parseGridTemplateAreasColumnNames(StringView gridRowNames)
-{
-    ASSERT(!gridRowNames.isEmpty());
-    Vector<String> columnNames;
-    StringBuilder areaName;
-    for (auto character : gridRowNames.codeUnits()) {
-        if (isCSSSpace(character)) {
-            if (!areaName.isEmpty()) {
-                columnNames.append(areaName.toString());
-                areaName.clear();
-            }
-            continue;
-        }
-        if (character == '.') {
-            if (areaName == "."_s)
-                continue;
-            if (!areaName.isEmpty()) {
-                columnNames.append(areaName.toString());
-                areaName.clear();
-            }
-        } else {
-            if (!isNameCodePoint(character))
-                return Vector<String>();
-            if (areaName == "."_s) {
-                columnNames.append(areaName.toString());
-                areaName.clear();
-            }
-        }
-        areaName.append(character);
-    }
-    if (!areaName.isEmpty())
-        columnNames.append(areaName.toString());
-    return columnNames;
-}
-
-bool parseGridTemplateAreasRow(StringView gridRowNames, NamedGridAreaMap& gridAreaMap, const size_t rowCount, size_t& columnCount)
-{
-    if (gridRowNames.containsOnly<isCSSSpace>())
-        return false;
-
-    auto columnNames = parseGridTemplateAreasColumnNames(gridRowNames);
-    if (!rowCount) {
-        columnCount = columnNames.size();
-        if (!columnCount)
-            return false;
-    } else if (columnCount != columnNames.size()) {
-        // The declaration is invalid if all the rows don't have the number of columns.
-        return false;
-    }
-
-    for (size_t currentColumn = 0; currentColumn < columnCount; ++currentColumn) {
-        const String& gridAreaName = columnNames[currentColumn];
-
-        // Unamed areas are always valid (we consider them to be 1x1).
-        if (gridAreaName == "."_s)
-            continue;
-
-        size_t lookAheadColumn = currentColumn + 1;
-        while (lookAheadColumn < columnCount && columnNames[lookAheadColumn] == gridAreaName)
-            lookAheadColumn++;
-
-        auto result = gridAreaMap.map.ensure(gridAreaName, [&] {
-            return GridArea(GridSpan::translatedDefiniteGridSpan(rowCount, rowCount + 1), GridSpan::translatedDefiniteGridSpan(currentColumn, lookAheadColumn));
-        });
-        if (!result.isNewEntry) {
-            auto& gridArea = result.iterator->value;
-
-            // The following checks test that the grid area is a single filled-in rectangle.
-            // 1. The new row is adjacent to the previously parsed row.
-            if (rowCount != gridArea.rows.endLine())
-                return false;
-
-            // 2. The new area starts at the same position as the previously parsed area.
-            if (currentColumn != gridArea.columns.startLine())
-                return false;
-
-            // 3. The new area ends at the same position as the previously parsed area.
-            if (lookAheadColumn != gridArea.columns.endLine())
-                return false;
-
-            gridArea.rows = GridSpan::translatedDefiniteGridSpan(gridArea.rows.startLine(), gridArea.rows.endLine() + 1);
-        }
-        currentColumn = lookAheadColumn - 1;
-    }
-
-    return true;
-}
-
-static RefPtr<CSSPrimitiveValue> consumeGridBreadth(CSSParserTokenRange& range, CSSParserMode mode)
-{
-    const CSSParserToken& token = range.peek();
-    if (isGridBreadthIdent(token.id()))
-        return consumeIdent(range);
-    if (token.type() == DimensionToken && token.unitType() == CSSUnitType::CSS_FR) {
-        if (range.peek().numericValue() < 0)
-            return nullptr;
-        return CSSPrimitiveValue::create(range.consumeIncludingWhitespace().numericValue(), CSSUnitType::CSS_FR);
-    }
-    return consumeLengthOrPercent(range, mode, ValueRange::NonNegative);
-}
-
-RefPtr<CSSValue> consumeGridTrackSize(CSSParserTokenRange& range, CSSParserMode mode)
-{
-    const CSSParserToken& token = range.peek();
-    if (identMatches<CSSValueAuto>(token.id()))
-        return consumeIdent(range);
-
-    if (token.functionId() == CSSValueMinmax) {
-        CSSParserTokenRange rangeCopy = range;
-        CSSParserTokenRange args = consumeFunction(rangeCopy);
-        auto minTrackBreadth = consumeGridBreadth(args, mode);
-        if (!minTrackBreadth || minTrackBreadth->isFlex() || !consumeCommaIncludingWhitespace(args))
-            return nullptr;
-        auto maxTrackBreadth = consumeGridBreadth(args, mode);
-        if (!maxTrackBreadth || !args.atEnd())
-            return nullptr;
-        range = rangeCopy;
-        return CSSFunctionValue::create(CSSValueMinmax, minTrackBreadth.releaseNonNull(), maxTrackBreadth.releaseNonNull());
-    }
-
-    if (token.functionId() == CSSValueFitContent)
-        return consumeFitContent(range, mode);
-
-    return consumeGridBreadth(range, mode);
-}
-
-// Appends to the passed in CSSGridLineNamesValue if any, otherwise creates a new one. Returns nullptr if an empty list is consumed.
-RefPtr<CSSGridLineNamesValue> consumeGridLineNames(CSSParserTokenRange& range, AllowEmpty allowEmpty)
-{
-    CSSParserTokenRange rangeCopy = range;
-    if (rangeCopy.consumeIncludingWhitespace().type() != LeftBracketToken)
-        return nullptr;
-
-    Vector<String, 4> lineNames;
-    while (auto lineName = consumeCustomIdentForGridLine(rangeCopy))
-        lineNames.append(lineName->customIdent());
-    if (rangeCopy.consumeIncludingWhitespace().type() != RightBracketToken)
-        return nullptr;
-    range = rangeCopy;
-    if (allowEmpty == AllowEmpty::No && lineNames.isEmpty())
-        return nullptr;
-    return CSSGridLineNamesValue::create(lineNames);
-}
-
-static bool consumeGridTrackRepeatFunction(CSSParserTokenRange& range, CSSParserMode mode, CSSValueListBuilder& list, bool& isAutoRepeat, bool& allTracksAreFixedSized)
-{
-    CSSParserTokenRange args = consumeFunction(range);
-
-    CSSValueListBuilder repeatedValues;
-
-    // The number of repetitions for <auto-repeat> is not important at parsing level
-    // because it will be computed later, let's set it to 1.
-    size_t repetitions = 1;
-    auto autoRepeatType = consumeIdentRaw<CSSValueAutoFill, CSSValueAutoFit>(args);
-    isAutoRepeat = autoRepeatType.has_value();
-    if (!isAutoRepeat) {
-        auto repetition = consumePositiveIntegerRaw(args);
-        if (!repetition)
-            return false;
-        repetitions = clampTo<size_t>(static_cast<size_t>(*repetition), 0, GridPosition::max());
-    }
-    if (!consumeCommaIncludingWhitespace(args))
-        return false;
-    if (auto lineNames = consumeGridLineNames(args))
-        repeatedValues.append(lineNames.releaseNonNull());
-
-    size_t numberOfTracks = 0;
-    while (!args.atEnd()) {
-        auto trackSize = consumeGridTrackSize(args, mode);
-        if (!trackSize)
-            return false;
-        if (allTracksAreFixedSized)
-            allTracksAreFixedSized = isGridTrackFixedSized(*trackSize);
-        repeatedValues.append(trackSize.releaseNonNull());
-        ++numberOfTracks;
-        if (auto lineNames = consumeGridLineNames(args))
-            repeatedValues.append(lineNames.releaseNonNull());
-    }
-    // We should have found at least one <track-size> or else it is not a valid <track-list>.
-    if (!numberOfTracks)
-        return false;
-
-    if (isAutoRepeat)
-        list.append(CSSGridAutoRepeatValue::create(*autoRepeatType, WTFMove(repeatedValues)));
-    else {
-        // We clamp the repetitions to a multiple of the repeat() track list's size, while staying below the max grid size.
-        repetitions = std::min(repetitions, GridPosition::max() / numberOfTracks);
-        list.append(CSSGridIntegerRepeatValue::create(repetitions, WTFMove(repeatedValues)));
-    }
-    return true;
-}
-
-static bool consumeSubgridNameRepeatFunction(CSSParserTokenRange& range, CSSValueListBuilder& list, bool& isAutoRepeat)
-{
-    CSSParserTokenRange args = consumeFunction(range);
-    size_t repetitions = 1;
-    isAutoRepeat = consumeIdentRaw<CSSValueAutoFill>(args).has_value();
-    if (!isAutoRepeat) {
-        auto repetition = consumePositiveIntegerRaw(args);
-        if (!repetition)
-            return false;
-        repetitions = clampTo<size_t>(static_cast<size_t>(*repetition), 0, GridPosition::max());
-    }
-    if (!consumeCommaIncludingWhitespace(args))
-        return false;
-
-    CSSValueListBuilder repeatedValues;
-    do {
-        auto lineNames = consumeGridLineNames(args, AllowEmpty::Yes);
-        if (!lineNames)
-            return false;
-        repeatedValues.append(lineNames.releaseNonNull());
-    } while (!args.atEnd());
-
-    if (isAutoRepeat)
-        list.append(CSSGridAutoRepeatValue::create(CSSValueAutoFill, WTFMove(repeatedValues)));
-    else
-        list.append(CSSGridIntegerRepeatValue::create(repetitions, WTFMove(repeatedValues)));
-    return true;
-}
-
-RefPtr<CSSValue> consumeGridTrackList(CSSParserTokenRange& range, const CSSParserContext& context, TrackListType trackListType)
-{
-    if (context.masonryEnabled && range.peek().id() == CSSValueMasonry)
-        return consumeIdent(range);
-    bool seenAutoRepeat = false;
-    if (trackListType == GridTemplate && range.peek().id() == CSSValueSubgrid) {
-        consumeIdent(range);
-        CSSValueListBuilder values;
-        while (!range.atEnd() && range.peek().type() != DelimiterToken) {
-            if (range.peek().functionId() == CSSValueRepeat) {
-                bool isAutoRepeat;
-                if (!consumeSubgridNameRepeatFunction(range, values, isAutoRepeat))
-                    return nullptr;
-                if (isAutoRepeat && seenAutoRepeat)
-                    return nullptr;
-                seenAutoRepeat = seenAutoRepeat || isAutoRepeat;
-            } else if (auto value = consumeGridLineNames(range, AllowEmpty::Yes))
-                values.append(value.releaseNonNull());
-            else
-                return nullptr;
-        }
-        return CSSSubgridValue::create(WTFMove(values));
-    }
-
-    bool allowGridLineNames = trackListType != GridAuto;
-    if (!allowGridLineNames && range.peek().type() == LeftBracketToken)
-        return nullptr;
-
-    CSSValueListBuilder values;
-    bool allowRepeat = trackListType == GridTemplate;
-    bool allTracksAreFixedSized = true;
-    if (auto lineNames = consumeGridLineNames(range))
-        values.append(lineNames.releaseNonNull());
-    do {
-        bool isAutoRepeat;
-        if (range.peek().functionId() == CSSValueRepeat) {
-            if (!allowRepeat)
-                return nullptr;
-            if (!consumeGridTrackRepeatFunction(range, context.mode, values, isAutoRepeat, allTracksAreFixedSized))
-                return nullptr;
-            if (isAutoRepeat && seenAutoRepeat)
-                return nullptr;
-            seenAutoRepeat = seenAutoRepeat || isAutoRepeat;
-        } else if (RefPtr<CSSValue> value = consumeGridTrackSize(range, context.mode)) {
-            if (allTracksAreFixedSized)
-                allTracksAreFixedSized = isGridTrackFixedSized(*value);
-            values.append(value.releaseNonNull());
-        } else
-            return nullptr;
-        if (seenAutoRepeat && !allTracksAreFixedSized)
-            return nullptr;
-        if (!allowGridLineNames && range.peek().type() == LeftBracketToken)
-            return nullptr;
-        if (auto lineNames = consumeGridLineNames(range))
-            values.append(lineNames.releaseNonNull());
-    } while (!range.atEnd() && range.peek().type() != DelimiterToken);
-    return CSSValueList::createSpaceSeparated(WTFMove(values));
-}
-
-RefPtr<CSSValue> consumeGridTemplatesRowsOrColumns(CSSParserTokenRange& range, const CSSParserContext& context)
-{
-    if (range.peek().id() == CSSValueNone)
-        return consumeIdent(range);
-    if (context.masonryEnabled && range.peek().id() == CSSValueMasonry)
-        return consumeIdent(range);
-    return consumeGridTrackList(range, context, GridTemplate);
-}
-
-RefPtr<CSSValue> consumeGridTemplateAreas(CSSParserTokenRange& range)
-{
-    if (range.peek().id() == CSSValueNone)
-        return consumeIdent(range);
-
-    NamedGridAreaMap map;
-    size_t rowCount = 0;
-    size_t columnCount = 0;
-    while (range.peek().type() == StringToken) {
-        if (!parseGridTemplateAreasRow(range.consumeIncludingWhitespace().value(), map, rowCount, columnCount))
-            return nullptr;
-        ++rowCount;
-    }
-    if (!rowCount)
-        return nullptr;
-    return CSSGridTemplateAreasValue::create(WTFMove(map), rowCount, columnCount);
-}
 
 RefPtr<CSSValue> consumeLineBoxContain(CSSParserTokenRange& range)
 {
